FIX: 새로고침 시 recoil 상태의 초기화로 인한 깜빡임 50% 해결
- recoil selector를 이용하여 로그인 상태와 사용자 아이디의 초기값을
  비동기 응답으로 할당
- 헤더에서 useEffect로 매번 비동기 통신을 하는 대신 app에서 한번만 수행
- recoil 값이 없는 동안은 렌더링하지 않도록 수정
- recoil의 초기화 문제는 다음 배포 때 로컬 스토리지에 저장하는 방식으로
  해결할 예정
